First rework of go9p
Using this code on Akaros I've learned a few things, and this is the first
set of changes from what I've learned. I hope this is ok. There are more coming.

Unpackdir now returns how many bytes it unpacked. I realized that
Dir.size+2 is probably fine most of the time but it confused
people so I made it like that and they stopped worrying.

There are two tests now. They test real world scenarios we've seen. One exposes
a problem with directory reading. I'm going to need to fix that issue.

ufs is now a package, and the ufs example uses it. This makes it possible
to include ufs and other servers in a single server, but also to write tests
that use ufs as part of the test.

I realize these are kind of intrusive changes but Lucho said they seemed ok.
Let me know if there's a problem.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>