CLeanup some of the dir management.
This is for some later code that seems to improve performance
and fixes some bugs.

gstat returns an error now.
unpackdir returns a Dir array, the remaining byte slice, an amount, and an error

The other code uses that stuff. This has been extensively used and tested so should
work fine.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>