Slightly change directory handling and fix a few problems.
The Sysif pointer can be nil, or not the expected type, so check both
things. This means that some errors can be returned from, e.g., Stat,
even if the underlying stat succeeded, because it does not return the right
type. Or, maybe, we should just fake it? not sure yet.

Also, we allow the client to pass an offset that is not what the server
thinks it sent back. This can happen if the client needs to move back when
it got an incomplete entry, which seems to happen.

go9p_test now tests these types of errors, as well as adding a
test for a nasty unpack bug that was in there from day one. I can't believe
we (almost) never hit it.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>