Implement remove; fix rwstat; fix table; allow absolute paths
Remove did not work, and in fact Rwstat did not work.

There was a missing entry in a the Fc*size tables for the size of Rwstat.
The sense of which table to use was reversed.

This was uncovered in the unit test now in go9p_test.go

We allow absolute paths for the rename now because
mv /x/y/giantdir /x/y/Giantdir
is at least a ten bilion times more efficient than
cp -r /x/y/giantdir /x/Giantdir
rm -rf /x/y/giantdir
if the size of giantdir is measured in gigabytes.
especialy if you have to do it over 9p. The 9p naming
restriction on rename is stupid. Let the providers decide.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>