More remove cleanups and ufs fix
Sometimes external users of the ufs package want to set the root.
This was not possible before. Now there is a Root string
in the ufs struct.

This also adds a test to make sure people can't walk out of the mount
point via ..

Clients should not pass paths containing .. to the server but the server
should protect against it anyway.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>