Get the ufs server to work with a Plan 9 terminal.
This set of changes is tested with a Plan 9 terminal and works fine.

I added the option of adding a user for people who want to set that. It's not required.
I also make -root= actually work, but you have to use ufs.New() now. That's probably sensible
since the ufs package has options of its own.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>