In the UFS server, a Wstat with no changes now means 'fsync' if the file is open.
This follows common interpretation of the 'no op' Wstat.

Here is what we do now: each server type (e.g. ufs) interprets the Wstat request and works out what to do.
This leads to code bloat to some extent, as each server type has the same code for looking at the request.

The other way is to have each type of server (e.g. ufs) supply a set of interfaces, one per type of Wstat variable, and have
the higher level decide whether to call interface functions based on what is set in the Wstat.

Each approach has its merits and problems. I decided to leave it as is for now because the argument for having the server
handle the Wstat directly is pretty strong.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>