Add FFindOp extension point
The new interface is an operation to find a child in a directory by
name, and it is used, if implemented, in the implementation of Walk in
*srv.Fsrv.

The reason for adding this is that sometimes, for synthetic filesystems,
you want a file to be created implicitly when walking to it. As an
example, I've added a toy file server, ipfs. Initially it is empty, but
if you walk to a host name, a corresponding file will be created
implicitly and its contents will be the addresses of that host.

	% cat /mnt/ip/github.com
	140.82.118.4
	140.82.118.3
	% cat /mnt/ip/foobar
	cat: /mnt/ip/foobar: No such file or directory

In general, this extension point is needed if one wants to lazy-load the
contents of a directory. Usually the FOpenOp extension point can be used
for lazy-loading, but that fails if one walks straight to a child
without opening its parent, e.g., "cat /mnt/fs/parent/child" instead of
"cd /mnt/fs/parent ; ls ; cat child".