p/srv/examples/ipfs: expire child if created more than minute ago
A second use case for the FFindOp interface is to expire nodes lazily.
Suppose you have a file that represents information that's valid for
just some amount of time. You can check whether it's still valid when
walking to it, and if it's not, you can remove it from the file server
or update its contents and metadata as needed.