p/srv: honor error returned by implementers of srv.FClunkOp
In the below, the test program is "9p -a localhost:5640 ls" running
against a file server that implements srv.FClunkOp and always returns a
non-nil error.

The resulting 9P dialog using commit b4200817 is

        ...
	2019/02/23 21:18:24 >>> 127.0.0.1:46196 Tclunk tag 0 fid 1
	2019/02/23 21:18:24 <<< 127.0.0.1:46196 Rclunk tag 0
	2019/02/23 21:18:24 >>> 127.0.0.1:46196 Twalk tag 0 fid 0 newfid 1
	2019/02/23 21:18:24 <<< 127.0.0.1:46196 Rerror tag 0 ename 'fid already in use' ecode 0
	2019/02/23 21:18:24 >>> 127.0.0.1:46196 Tclunk tag 0 fid 0
	2019/02/23 21:18:24 <<< 127.0.0.1:46196 Rclunk tag 0

IIUC this is because the lack of return after the RespondError() made it
possible to overwrite the response buffer before the send goroutine
could send it, so that the response is in fact Rclunk. (Because of
another bug that will be addressed in another commit, the fid ref count
won't be decremented, which will result in the "fid already in use"
error.)

The same test after this commit gives this 9P dialog fragment:

	2019/02/23 21:29:51 >>> 127.0.0.1:46220 Tclunk tag 0 fid 1
	2019/02/23 21:29:51 <<< 127.0.0.1:46220 Rerror tag 0 ename 'synthetic error' ecode 0

(The same happens if one introduces a time.Sleep(time.Second) call after
the RespondError() call, giving the send goroutine a chance to send the
error.)